refactor: login, signup 페이지 label 제거 및 컴포넌트 index.ts로 모으기
