+++ components/common/Button.tsx
@@ -8,7 +8,8 @@ type Size =
   | 'xsmall'
   | 'small'
   | 'sign'
-  | 'responsive';
+  | 'responsive'
+  | 'modal';
 
 interface ButtonProps {
   variant: Variant;
@@ -70,6 +71,10 @@ function Button({
     case 'responsive': {
       combinedClassName +=
         ' desktop:py-7pxr desktop:px-29pxr desktop:text-14pxr tablet:py-6pxr tablet:px-23pxr tablet:text-14pxr mobile:py-7pxr mobile:px-37pxr mobile:text-12pxr';
+    }
+    case 'modal': {
+      combinedClassName +=
+        ' text-16pxr w-120pxr h-48pxr py-14pxr px-14pxr mobile:text-14pxr mobile:w-138pxr mobile:h-42pxr mobile:py-12pxr mobile:px-12pxr';
     }
     case 'small': {
       combinedClassName += ' py-7pxr px-29pxr text-14pxr';
