+++ components/index.ts
@@ -1,4 +1,9 @@
-export { default as ColorChips } from './common/ColorChips';
+export { default as SignHeader } from './sign/header/SignHeader';
+export { default as SignFooter } from './sign/footer/SignFooter';
+export { default as LoginForm } from './sign/form/LoginForm';
+export { default as SignUpForm } from './sign/form/SignUpForm';
 export { default as Button } from './common/Button';
+export { default as ColorChips } from './common/ColorChips';
 export { default as Input } from './common/Input';
-export { default as PasswordInput} from './common/PasswordInput'
+export { default as PasswordInput } from './common/PasswordInput';
+export { default as Modal } from './modal/Modal';
