+++ pages/signup/index.tsx
@@ -1,5 +1,4 @@
-import { SignFooter, SignHeader } from '@/components/sign';
-import SignUpForm from '@/components/sign/form/SignUpForm';
+import { SignFooter, SignHeader, SignUpForm } from '@/components/index';
 import SignLayout from '@/page-layout/SignLayout';
 
 export default function SignUpPage() {
